# Make Int{Map,Set} folds friendlier to optimizations