@meooow25 2025-06-21
Apparently SpecConstr does not trigger for `IntSet` in the `maximum` benchmarks like it does for `IntMap`. I should have double checked. But I don't understand why, so I've opened [GHC #26141](https://gitlab.haskell.org/ghc/ghc/-/issues/26141) about this.